database_jkm_wyx_yzh/requirements.txt
Django==4.2.5
django_ckeditor==6.7.0
Pillow==10.0.1
PyMySQL==1.1.0
